Cast to DOUBLE, not FLOAT, across the output layer (MLB-10)
Snowflake's FLOAT is 64-bit; DuckDB's is FLOAT4, 32-bit. So every
CAST(x AS FLOAT) in the output layer silently lost precision on DuckDB:
10.7 rendered as 10.699999809265137. Phase 0 fixed this class inside
dbt_league and output/ was never swept -- 161 sites across five files,
all one spelling.

DOUBLE is 64-bit on both engines and Snowflake treats it as a synonym for
FLOAT, verified before the sweep: FLOAT and DOUBLE agree on every
standings row, 0 mismatches.

This was the systemic cause behind most of the render A/B. Before it, 13
of 19 almanac tabs differed from the Snowflake goldens in what looked
like three separate classes -- wrong values, off-by-one counts, and whole
ROWS in a different order. All three were this: float32 sums landing just
under an integer, and perturbing sort keys enough to flip near-ties.
Fixing the root collapsed all 13 at once, and the ESPN almanac now
renders byte-identically off DuckDB, all 19 tabs.

Warehouse suite green 17/17, so the Snowflake side is unmoved.

+++ output/generate_season_report.py
@@ -347,7 +347,7 @@ def get_season_wasted_players(season_year, limit=5):
         WITH bench AS (
             SELECT player_id,
                    MAX(player_name) AS player_name,
-                   CAST(SUM(CAST(calculated_points AS DECIMAL(18, 6))) AS FLOAT) AS bench_wasted_pts,
+                   CAST(SUM(CAST(calculated_points AS DECIMAL(18, 6))) AS DOUBLE) AS bench_wasted_pts,
                    MAX(team_name) AS bench_team_name
             FROM fct_player_weekly_inactive_performance
             WHERE season_year = %s AND wasted_bucket = 'ROSTERED_INACTIVE'
@@ -357,8 +357,8 @@ def get_season_wasted_players(season_year, limit=5):
         active AS (
             SELECT player_id,
                    MAX_BY(team_name, matchup_period) AS active_team_name,
-                   CAST(SUM(CAST(GREATEST(0, -platform_points) AS DECIMAL(18, 6))) AS FLOAT) AS negative_active_pts,
-                   CAST(SUM(CAST(platform_points AS DECIMAL(18, 6))) AS FLOAT) AS active_points
+                   CAST(SUM(CAST(GREATEST(0, -platform_points) AS DECIMAL(18, 6))) AS DOUBLE) AS negative_active_pts,
+                   CAST(SUM(CAST(platform_points AS DECIMAL(18, 6))) AS DOUBLE) AS active_points
             FROM fct_player_weekly_active_performance
             WHERE season_year = %s
               AND {league_predicate()}
@@ -406,14 +406,14 @@ def get_team_wasted_totals(season_year):
     here too."""
     return query_snowflake(f"""
         WITH bench AS (
-            SELECT team_id, CAST(SUM(CAST(calculated_points AS DECIMAL(18, 6))) AS FLOAT) AS bench_pts
+            SELECT team_id, CAST(SUM(CAST(calculated_points AS DECIMAL(18, 6))) AS DOUBLE) AS bench_pts
             FROM fct_team_weekly_inactive_performance
             WHERE season_year = %s AND team_id IS NOT NULL
               AND {league_predicate()}
             GROUP BY team_id
         ),
         neg AS (
-            SELECT team_id, CAST(SUM(CAST(GREATEST(0, -platform_points) AS DECIMAL(18, 6))) AS FLOAT) AS neg_pts
+            SELECT team_id, CAST(SUM(CAST(GREATEST(0, -platform_points) AS DECIMAL(18, 6))) AS DOUBLE) AS neg_pts
             FROM fct_player_weekly_active_performance
             WHERE season_year = %s
               AND {league_predicate()}
